% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_caschool.R
\docType{data}
\name{df_opt_caschool}
\alias{df_opt_caschool}
\title{Test Dataset California Test Score Data}
\format{csv}
\source{
\url{https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des}
}
\usage{
data(df_opt_caschool)
}
\description{
420 school level observations a cross-section from 1998-1999
distcod, district code
county, county
district, district
grspan, grade span of district
enrltot, total enrollment
teachers, number of teachers
calwpct, percent qualifying for CalWorks
mealpct, percent qualifying for reduced-price lunch
computer, number of computers
testscr, average test score (read.scr+math.scr)/2
compstu, computer per student
expnstu, expenditure per student
str, student teacher ratio
avginc, district average income
elpct, percent of English learners
readscr, average reading score
mathscr, average math score
}
\examples{
data(df_opt_caschool)
head(df_opt_caschool, 10)
}
\references{
Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.
}
\keyword{datasets}
