% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_dtgch_cbem4_rrlop.R
\docType{data}
\name{df_opt_dtgch_cbem4_rrlop}
\alias{df_opt_dtgch_cbem4_rrlop}
\title{Optimal Linear Allocation based on dtgch cbem4 solved by solin relow}
\format{csv}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1570677X16300107}
}
\usage{
data(df_opt_dtgch_cbem4_rrlop)
}
\description{

}
\examples{
data(df_opt_dtgch_cbem4_rrlop)
ar_opti_inpalc <- df_opt_dtgch_cbem4_rrlop[['opti_allocate']]
ar_opti_expout <- df_opt_dtgch_cbem4_rrlop[['opti_exp_outcome']]
ar_opti_inpalc_cv <- (ar_opti_inpalc - mean(ar_opti_inpalc))/sd(ar_opti_inpalc)
ar_opti_expout_cv <- (ar_opti_expout - mean(ar_opti_expout))/sd(ar_opti_expout)
# Print
head(df_opt_dtgch_cbem4_rrlop, 10)
summary(df_opt_dtgch_cbem4_rrlop)
print('opti allocation and outcomes rescaled:')
print(cbind(ar_opti_expout_cv, ar_opti_inpalc_cv))
}
\references{
“Early life height and weight production functions with endogenous energy and protein inputs”, September 2016, 65-81, Economics & Human Biology, 22, Esteban Puentes, Fan Wang, Jere R. Behrman, Flavio Cunha, John Hoddinott, John A. Maluccio, Linda S. Adair, Judith Borja and Reynaldo Martorell
}
\keyword{datasets}
