% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_dtgch_cbem4_rrlop_allrh.R
\docType{data}
\name{df_opt_dtgch_cbem4_rrlop_allrh}
\alias{df_opt_dtgch_cbem4_rrlop_allrh}
\title{Optimal Linear Allocation based on dtgch cbem4 solved by solin relow, multiple rhos, collected}
\format{csv}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1570677X16300107}
}
\usage{
data(df_opt_dtgch_cbem4_rrlop_allrh)
}
\description{

}
\examples{
data(df_opt_dtgch_cbem4_rrlop_allrh)
# Print
head(df_opt_dtgch_cbem4_rrlop_allrh, 10)
head(df_opt_dtgch_cbem4_rrlop_allrh[,10:17], 1100)
}
\references{
“Early life height and weight production functions with endogenous energy and protein inputs”, September 2016, 65-81, Economics & Human Biology, 22, Esteban Puentes, Fan Wang, Jere R. Behrman, Flavio Cunha, John Hoddinott, John A. Maluccio, Linda S. Adair, Judith Borja and Reynaldo Martorell
}
\keyword{datasets}
