% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_dtgch_cbem4_rrlop_argin.R
\docType{data}
\name{df_opt_dtgch_cbem4_rrlop_argin}
\alias{df_opt_dtgch_cbem4_rrlop_argin}
\title{Optimal Linear Allocation dtgch cbem4, solin relow, multiple rhos, GINI}
\format{csv}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1570677X16300107}
}
\usage{
data(df_opt_dtgch_cbem4_rrlop_argin)
}
\description{
Generated by *\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_solin_relow.html}*, opti allocate and expected outcome
}
\examples{
data(df_opt_dtgch_cbem4_rrlop_argin)
print(as.matrix(df_opt_dtgch_cbem4_rrlop_argin))
}
\references{
“Early life height and weight production functions with endogenous energy and protein inputs”, September 2016, 65-81, Economics & Human Biology, 22, Esteban Puentes, Fan Wang, Jere R. Behrman, Flavio Cunha, John Hoddinott, John A. Maluccio, Linda S. Adair, Judith Borja and Reynaldo Martorell
}
\keyword{datasets}
