% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_lalonde_training.R
\docType{data}
\name{df_opt_lalonde_training}
\alias{df_opt_lalonde_training}
\title{Test Dataset Job Training Effects Lalonde AER 1986}
\format{
csv
}
\source{
\url{https://www.jstor.org/stable/1806062?}
}
\usage{
data(df_opt_lalonde_training)
}
\description{
trt, a numeric vector identifying the study in which the subjects were enrolled (0 = Control, 1 = treated). Male module, 297 treated. 425 untreated.
age, age (in years).
educ, years of education.
black, (0 = not black, 1 = black).
hisp, (0 = not hispanic, 1 = hispanic).
marr, (0 = not married, 1 = married).
nodeg, (0 = completed high school, 1 = dropout).
re74, real earnings in 1974.
re75, real earnings in 1975.
re78, real earnings in 1978.
}
\examples{
data(df_opt_lalonde_training)
head(df_opt_lalonde_training, 10)
}
\references{
Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.
}
\keyword{datasets}
