% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_opt_sodis_val_nsw_stimulus.R
\name{ffp_nsw_opt_sodis_value}
\alias{ffp_nsw_opt_sodis_value}
\title{Value at each W point along Queue, for one planner rho, given optimal
allocation. NSW version.}
\usage{
ffp_nsw_opt_sodis_value(
  fl_rho,
  df_queue_il,
  bl_return_allQ_V = FALSE,
  bl_return_inner_V = FALSE,
  svr_id_i = "id_i",
  svr_id_il = "id_il",
  svr_D_il = "D_il",
  svr_inpalc = "Q_il",
  svr_D_Wbin_il = "D_Wbin_il",
  svr_A_il = "A_il",
  svr_alpha_il = "alpha_il",
  svr_beta_i = "beta_i",
  svr_measure_i = NA,
  svr_betameasure_i = NA,
  svr_V_cumu_l = "V_sum_l",
  svr_V_inner_Q_il = "V_inner_Q_il",
  svr_V_star_Q_il = "V_star_Q_il"
)
}
\arguments{
\item{bl_return_allQ_V}{boolean if true returns value along the entire
allocation queue including queue segments after resource limits. This is
more time consuming when set to true.}

\item{svr_beta_i}{string variable name for planner bias}

\item{svr_measure_i}{string NO LONGER USED variable name for mass for this
type of recipient, default NA mass of recipient is the measure of
recipient of this type in the population. This measure does not impact
relative ranking optimal allocation across types, but determines how much
to push individual types further along the allocation queue back. this
should be 'mass_i', representing the measure of individuals in this group}

\item{svr_betameasure_i}{string variable name for the weight considerating
both mass and beta jointly, can not simply multiply beta and measure
together. Suppose two people, beta is 0.5 for each person. And then we
have 0.5 mass for each person, simply multiplying the two together
genreates 0.5x0.5 vs 0.5x0.5, don't sum up to the existing total mass
even.}
}
\description{
After solving for optimal allocating queue, at each Q point,
  which corresponds to different level of aggregate resources available,
  show the value given optimal choices this is an array of values, at each
  Q point. Potentially output for all Q. Or only show this for all queue
  points up to the actual resource limit, D_Wbin_il = 1.

  This version of the function is adpated specially for
  \href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3839890}{Nygaard,
  Sorensen, and Wang (2021)} for the optimal allocation of stimulus checks
  paper, and is based on \code{\link{ffp_opt_sodis_value()}}.
}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/reference/ffp_opt_sodis_value.html}
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
