% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_opt_anlyz.R
\name{ffp_opt_anlyz_rhgin}
\alias{ffp_opt_anlyz_rhgin}
\title{Solves for optimal allocation problem over many rhos, and computes gini}
\usage{
ffp_opt_anlyz_rhgin(
  df,
  svr_id_i,
  svr_A_i,
  svr_alpha_i,
  svr_beta_i,
  fl_N_agg,
  ar_rho,
  svr_inpalc = "optiallocate",
  svr_expout = "opti_exp_outcome"
)
}
\arguments{
\item{df}{tibble data table including variables using svr names below each row is potentially an individual who will receive alternative allocations}

\item{svr_id_i}{string name of the identify key for each i, to make sure merging happens properly}

\item{svr_A_i}{string name of the A_i variable, dot product of covariates and coefficients}

\item{svr_alpha_i}{string name of the alpha_i variable, individual specific elasticity information}

\item{svr_beta_i}{string name of the beta_i variable, relative preference weight for each child}

\item{fl_N_agg}{float total resource avaible for allocation, if not specific, sum by svr_N_i}

\item{ar_rho}{array preferences for equality for the planner}

\item{svr_inpalc}{string variable name for newly generated input optimal allocation}

\item{svr_expout}{string variable name for newly generated expected outcome}
}
\value{
a dataframe that expands the df inputs with additional results.

a list with a dataframe and an array
\itemize{
  \item df_opti_alloc_all_rho - table with columns for optimal allocation and exp outcome different rhos
  \item mt_opti_alloc_all_rho - a matrix row = indi, col = rho, value = optimal allocation
  \item mt_expc_outcm_all_rho - a matrix row = indi, col = rho, value = exp outcome optimal allocations
  \item mt_gini - a matrix row = rhos, col = opti-allo and exp-out, value = gini
}
}
\description{
Works with linear, log linear and discrete allocation problems. The function
invokes ffp_opt_solin_relow, has parameters that are also required for that function.
The addition here is *fl_rho* becomes *ar_rho*
}
\examples{
data(df_opt_dtgch_cbem4)
df <- df_opt_dtgch_cbem4
svr_id_i <- 'indi.id'
svr_A_i <- 'A_lin'
svr_alpha_i <- 'alpha_lin'
svr_beta_i <- 'beta'
fl_N_agg <- 10000
ar_rho = c(-50, -10, -0.1, 0.1, 0.5, 0.7)
ls_lin_solu_all_rhos <- ffp_opt_anlyz_rhgin(df, svr_id_i, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, ar_rho)
df_opti_alloc_all_rho <- ls_lin_solu_all_rhos$df_opti_alloc_all_rho
summary(df_opti_alloc_all_rho)
mt_opti_alloc_all_rho <- ls_lin_solu_all_rhos$mt_opti_alloc_all_rho
summary(mt_opti_alloc_all_rho)
mt_expc_outcm_all_rho <- ls_lin_solu_all_rhos$mt_expc_outcm_all_rho
summary(mt_expc_outcm_all_rho)
mt_gini <- ls_lin_solu_all_rhos$mt_gini
summary(mt_gini)
}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/reference/ffp_opt_anlyz_rhgin.html}
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_anlyz_rhgin.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
