% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_opt_sodis.R
\name{ffp_opt_sodis_rev}
\alias{ffp_opt_sodis_rev}
\title{Discrete Problem Resource Equivalent Variation}
\usage{
ffp_opt_sodis_rev(
  fl_rho,
  it_w_agg,
  df_input_ib,
  df_queue_il_with_V,
  svr_A_i_l0 = "A_i_l0",
  svr_alpha_o_i = "alpha_o_i",
  svr_inpalc = "Q_il",
  svr_beta_i = "beta_i",
  svr_V_star_Q_il = "V_star_Q_il"
)
}
\description{
The function uses the Value already computed along the Queue
}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/reference/ffp_opt_sodis_rev.html}
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
