% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_welfarechecks_graph.R
\name{ffp_snw_graph_optimal}
\alias{ffp_snw_graph_optimal}
\title{Nygaard, Sorernsen and Wang (2020) Optimal Allocation by household type graphs.}
\usage{
ffp_snw_graph_optimal(
  ar_rho,
  df_input_il_noninc_covar,
  df_alloc_i_long_covar_c,
  df_alloc_i_long_covar_v,
  ls_st_gen_imgs = c("mass", "mc", "checks_c", "mv", "checks_v"),
  ls_st_save_imgs = c("mass", "mc", "checks_c", "mv", "checks_v"),
  st_file_type = "",
  slb_add_title = "",
  stg_subtitle = "",
  stg_caption = "",
  st_img_suffix = "",
  bl_save_img = FALSE,
  it_img_width = 135,
  it_img_height = 96,
  st_img_units = "mm",
  it_img_res = 300,
  it_img_pointsize = 7,
 
    spt_img_save = "C:/Users/fan/Documents/Dropbox (UH-ECON)/PrjNygaardSorensenWang/Results/2020-07-29/Graphs/"
)
}
\description{
SNW optimal Graphs
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
