% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_stimulus_bush_2008.R
\name{ffp_snw_stimulus_checks_bush}
\alias{ffp_snw_stimulus_checks_bush}
\title{US Economic Recovery Act of 2008 Bush Stimulus Check amounts (tax-rebates)
by household type and income array for Nygaard, Sorernsen and Wang (2021)}
\usage{
ffp_snw_stimulus_checks_bush(
  ar_income = seq(0, 250000, 5),
  it_kids = 0,
  bl_marital = 0,
  fl_stimulus_child = 300,
  fl_stimulus_adult_min = 300,
  fl_stimulus_adult_max = 600,
  fl_per_adult_phase_out = 75000,
  fl_phase_out_per_dollar_income = 0.05
)
}
\arguments{
\item{ar_income}{array of income points over which to evaluate taxable
income and also tax liability by kids count and marital status.}

\item{it_kids}{integer the number of children from 0 to 4 allowed}

\item{bl_marital}{boolean if 1 means married, if 0 means unmarried}

\item{fl_stimulus_child}{float the amount of stimulus per child, note that
this is unrelated to the tax-liability issue.}

\item{fl_stimulus_adult_min}{float the amount of minimum stimulus per adult
(for lower income households), even if tax liability is below this, so no
tax-rebate should be given, still provide this stimulus amount.}

\item{fl_stimulus_adult_max}{float the amount of max stimulus per adult, in
another word, the maximum amount of tax-rebate.}

\item{fl_per_adult_phase_out}{float the amount per-adult (household-head
only or household-head plus spouse) where phase-out for
stimulus/tax-rebate begin going down. 75k under the actual policy for
singles, and 150k for married.}

\item{fl_phase_out_per_dollar_income}{for every addition dollar beyond the
phaseout point, how much stimulus/tax-rebate to reduce. 0.05 means for
every additional dollar of income, stimulus go down by 5 cents.}
}
\value{
an array of tax-liabilities for particular kids count and martial
  status along an array of income levels
}
\description{
Bush checks, not used in matlab, meaning matlab solution
  functions do not call this. There are separate R functions that implement
  the checks that are unrelated to this, this is just to graph out the
  stimulus schedule for visualization

  Economic Stimulus Act of 2008: 600 ($1,200) for singles (couples) making
  less than $75,000 ($150,000). $300 per child. Amount phases out at a rate
  of 5 percent (drops by $50 per $1,000 in income exceeding $75,000
  ($150,000)).

  IRS rules for stimulus checks (tax rebates) under the Economic Stimulus
  Act of 2008:
  \url{https://www.irs.gov/newsroom/single-head-of-household-with-children},
   \url{https://www.irs.gov/newsroom/married-with-children},
  \url{https://www.irs.gov/newsroom/married-without-qualifying-children},
  and
  \url{https://www.irs.gov/newsroom/single-without-qualifying-children}.
}
\examples{
ar_income <- c(1e4, 2e4, 4e4, 8e4, 1.6e5)
it_kids <- 0
bl_marital <- 0
ar_stimulus_check <- ffp_snw_stimulus_checks_bush(ar_income, it_kids, bl_marital)
print('ar_stimulus_check')
print(ar_stimulus_check)

}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_snw_stimulus_bush_2008.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
