% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_stimulus_bush_2008.R
\name{ffp_snw_stimulus_checks_bush_add2dfid}
\alias{ffp_snw_stimulus_checks_bush_add2dfid}
\title{Given dataframe solve for stimulus across household types and find relaxed
policy bounds as defined by counterfactuals.}
\usage{
ffp_snw_stimulus_checks_bush_add2dfid(
  df_id,
  it_income_n_in_seg = 100,
  fl_multiple = 54831,
  fl_percheck_dollar = 100,
  fl_stimulus_child = 300,
  fl_stimulus_adult_min = 300,
  fl_stimulus_adult_max = 600,
  fl_per_adult_phase_out = 75000,
  fl_phase_out_per_dollar_income = 0.05
)
}
\arguments{
\item{df_id}{dataframe the dataframe that is generated at the start of the
allocation problem, it contains the individual id, marital status, kids
count, and a string varaible indicating lower and upper bound for current
household type's income bin.}

\item{it_income_n_in_seg}{integer the number of points between min and max
income bounds for the current income bin at which to evaluate stimulus,
and to average.}

\item{fl_multiple}{integer what is 1 unit in dollar in 2008, equal to 54831,
median income.}

\item{fl_percheck_dollar}{integer we discretize the stimulus problem, how
much is each stimulus check in dollars?}

\item{fl_stimulus_child}{float the amount of stimulus per child, note that
this is unrelated to the tax-liability issue.}

\item{fl_stimulus_adult_min}{float the amount of minimum stimulus per adult
(for lower income households), even if tax liability is below this, so no
tax-rebate should be given, still provide this stimulus amount.}

\item{fl_stimulus_adult_max}{float the amount of max stimulus per adult, in
another word, the maximum amount of tax-rebate.}

\item{fl_per_adult_phase_out}{float the amount per-adult (household-head
only or household-head plus spouse) where phase-out for
stimulus/tax-rebate begin going down. 75k under the actual policy for
singles, and 150k for married.}

\item{fl_phase_out_per_dollar_income}{for every addition dollar beyond the
phaseout point, how much stimulus/tax-rebate to reduce. 0.05 means for
every additional dollar of income, stimulus go down by 5 cents.}
}
\value{
an array of tax-liabilities for particular kids count and martial
 status along an array of income levels
}
\description{
Given dataframe solve for US Economic Recovery Act of 2008 Bush
 Stimulus Check amounts (tax-rebates) by household type and income array
 for Nygaard, Sorernsen and Wang (2021) for each household type. Also
 compute maximum allocation bounds based on tax-liability to consider
 alternatives to the actual stimulus (tax-rebate) policy.

 We have a dataframe of households, where each household is defined by the
 number of kids in the household, marital status, and also income bin. Note
 that this is an income bin, not a specific income level. We computes an
 approximate income-bin (and marital status and kids count) specific
 stimulus amount by evaluating the stimulus checks function along a fine
 grid of income levels from the min to the max point of the income-bin, and
 simply take the average.

 We do this first for the actual stimulus that households should receive
 under the Economic Stimulus Act of 2008. We then adjust parameters for the
 stimulus function and compute alternative max-stimulus bounds for each
 income bin.
}
\examples{
data(df_nsw_tiny_chk168_df_id)
df_id <- df_nsw_tiny_chk168_df_id
it_income_n_in_seg <- 100
df_id_checkadded_actual <- ffp_snw_stimulus_checks_bush_add2dfid(df_id, it_income_n_in_seg)
print('df_id_checkadded_actual')
print(df_id_checkadded_actual)

}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_snw_stimulus_bush_2008.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
