% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_stimulus_bush_2008.R
\name{ffp_snw_tax_liability}
\alias{ffp_snw_tax_liability}
\title{US 2008 Tax liability by household type given income for Nygaard, Sorernsen
and Wang (2021)}
\usage{
ffp_snw_tax_liability(
  ar_income = seq(0, 20000, length.out = 5),
  bl_verbose = FALSE
)
}
\arguments{
\item{ar_income}{array of income points over which to evaluate taxable
income and also tax liability by kids count and marital status.}
}
\value{
a list of 3-d arrays \itemize{ \item mn_taxable_income - 2008
  taxable income given income 3d array where 1st dimension is income, 2nd
  dimension is marital status, 3rd dimension is kids count \item
  mn_tax_liability - 2008 taxable liability given income 3d array }
}
\description{
We can study the effects of the 2008 Tax Rebate. The Tax
  rebate is a rebate based on how much tax was paid, so we need to know
  taxable income and tax liability. These differ by income, household
  marital status, and the count of children.

  Given an array of pre-tax income values, we compute for from 0 to 4 kids
  and both married and unmarried taxable-income and tax-liability at all
  points along the income array.

  Deductions are from 2008 income
  \url{https://www.irs.gov/pub/irs-prior/f1040--2008.pdf}. Tax brackets
  from 2008 are here
  \url{https://www.irs.gov/pub/irs-prior/i1040tt--2008.pdf}.
}
\examples{
ar_income <- c(1e4, 2e4, 4e4, 8e4, 1.6e5)
ls_taxable <- ffp_snw_tax_liability(ar_income)
mn_taxable_income <- ls_taxable$mn_taxable_income
mn_tax_liability <- ls_taxable$mn_tax_liability
print('mn_taxable_income')
print(mn_taxable_income)
print('mn_tax_liability')
print(mn_tax_liability)

}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_snw_stimulus_bush_2008.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
