% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffy_opt_dtgch.R
\name{ffy_opt_dtgch_cbem4}
\alias{ffy_opt_dtgch_cbem4}
\title{Generates A and alpha for linear and log linear problems from Guat Cebu 4 Groups}
\usage{
ffy_opt_dtgch_cbem4()
}
\value{
a list with two tibble with guatemala and cebu dataset only some columns and A and alpha lin loglin
\itemize{
  \item df_raw - Dataframe from Guat/Cebu subsetted Ceb only, 4 categories
  \item df_esti - A dataframe with 5 columns, lin and loglin A, alpha and beta.
}
}
\description{
Given a dataset with Y, x1, xothers. Regress with log linear and linear structures
to obtain A and alpha that the linear and log linear optimal allocation problems understand.
dtgch = data guatemala cebu height. There discrete groups, allowing for alpha to differ.
}
\examples{
ls_opti_alpha_A <- ffy_opt_dtgch_cbem4()
head(ls_opti_alpha_A$df_raw, 10)
head(ls_opti_alpha_A$df_esti, 10)
}
\references{
\url{https://fanwangecon.github.io/PrjOptiAlloc/reference/ffy_opt_dtgch_fourg.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
