% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffy_opt_ghthm.R
\name{ffy_opt_ghthm_dk}
\alias{ffy_opt_ghthm_dk}
\title{ghthm = graph theme, support theme files}
\usage{
ffy_opt_ghthm_dk()
}
\value{
ggplot theme
}
\description{
Theme from Dominik Koch found on luisdva.github.io
}
\references{
\url{https://luisdva.github.io/rstats/dog-bump-chart/}
}
\author{
Dominik Koch, \url{https://luisdva.github.io/rstats/dog-bump-chart/}
}
