% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffy_opt_solin.R
\name{ffy_opt_solin_flinr}
\alias{ffy_opt_solin_flinr}
\title{Solves the linear allocation problem with init condi A, marg. prod alpha, and planner preferences}
\usage{
ffy_opt_solin_flinr(df, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, fl_rho)
}
\arguments{
\item{df}{tibble data table including variables using svr names below each row is potentially an individual who will receive alternative allocations}

\item{svr_A_i}{string name of the A_i variable, dot product of covariates and coefficients}

\item{svr_alpha_i}{string name of the alpha_i variable, individual specific elasticity information}

\item{svr_beta_i}{string name of the beta_i variable, relative preference weight for each child}

\item{fl_N_agg}{float total resource avaible for allocation, if not specific, sum by svr_N_i}

\item{fl_rho}{float preference for equality for the planner}
}
\value{
a dataframe that expands the df inputs with additional results.

a list with a dataframe and an array
\itemize{
  \item df_opti - Dataframe with various statistcs related to optimal allocation, all intermediary stats
  \item ar_opti - Array where each element is an optimal choice for each individual
}
}
\description{
This is the solution to the linear optimal allocation problem. Relative allocations, summed splined, inversion.
}
\examples{
data(df_opt_dtgch_cbem4)
df <- df_opt_dtgch_cbem4
svr_A_i <- 'A_lin'
svr_alpha_i <- 'alpha_lin'
svr_beta_i <- 'beta'
fl_N_agg <- 10000
fl_rho <- -30
ls_lin_solu <- ffy_opt_solin_flinr(df, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, fl_rho)
df_opti <- ls_lin_solu$df_opti
ar_opti <- ls_lin_solu$ar_opti
summary(df_opti)
print(ar_opti)
}
\references{
\url{https://fanwangecon.github.io/REconTools/reference/ffy_opt_solin_flinr.html}
\url{https://fanwangecon.github.io/REconTools/articles/ffy_opt_solin_flinr.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
