% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_opt_birthwt.R
\docType{data}
\name{df_opt_birthwt}
\alias{df_opt_birthwt}
\title{Test Dataset Birth Weight}
\format{csv}
\source{
\url{https://onlinelibrary.wiley.com/doi/book/10.1002/9781118548387}
}
\usage{
data(df_opt_birthwt)
}
\description{
Dataset from Baystate Medical Center in Springfield, Massachusetts (Hosmer et. al. 1998)
low: indicator of birth weight less than 2.5 kg.
age: mother's age in years.
lwt: mother's weight in pounds at last menstrual period.
race: mother's race (1 = white, 2 = black, 3 = other).
smoke: smoking status during pregnancy.
ptl: number of previous premature labours.
ht: history of hypertension.
ui: presence of uterine irritability.
ftv: number of physician visits during the first trimester.
bwt: birth weight in grams.
}
\examples{
data(df_opt_birthwt)
head(df_opt_birthwt, 10)
}
\references{
HOSMER, D.W., HOSMER, T., LE CESSIE, S. and LEMESHOW, S. (1997), A COMPARISON OF GOODNESS‐OF‐FIT TESTS FOR THE LOGISTIC REGRESSION MODEL. Statist. Med., 16: 965-980. doi:10.1002/(SICI)1097-0258(19970515)16:9<965::AID-SIM509>3.0.CO;2-O
}
\keyword{datasets}
