% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_welfarechecks_graph.R
\name{ffp_snw_graph_feasible}
\alias{ffp_snw_graph_feasible}
\title{Feasible Graphs considering ages}
\usage{
ffp_snw_graph_feasible(
  ar_rho,
  df_input_il_noninc_covar,
  df_alloc_i_long_covar_c,
  df_alloc_i_long_covar_v,
  ls_st_gen_imgs = c("mass", "mc", "checks_c", "mv", "mlogc", "checks_v", "checks_cjv"),
  ls_st_save_imgs = c("mass", "mc", "checks_c", "mv", "mlogc", "checks_v",
    "checks_cjv"),
  st_file_type = "",
  slb_add_title = "",
  stg_subtitle = "",
  stg_caption = "",
  bl_no_title_caption = TRUE,
  bl_optimal_with_age_groups = FALSE,
  st_img_suffix = "",
  bl_save_img = FALSE,
  it_img_width = 135,
  it_img_height = 96,
  st_img_units = "mm",
  it_img_res = 300,
  it_img_pointsize = 7,
 
    spt_img_save = "C:/Users/fan/Documents/Dropbox (UH-ECON)/PrjNygaardSorensenWang/Results/2020-07-29/Graphs/"
)
}
\description{
SNW feasible Graphs considering ages
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
