% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_snw_welfarechecks_input.R
\name{ffp_snw_process_inputs}
\alias{ffp_snw_process_inputs}
\title{Processing Matlab Simulation results for allocation for Nygaard, Sorernsen and Wang (2020)}
\usage{
ffp_snw_process_inputs(
 
    srt_simu_path = "C:/Users/fan/Documents/Dropbox (UH-ECON)/PrjNygaardSorensenWang/Output/",
  snm_simu_csv = "snwx_v_planner_moredense_a100zh266_e2m2.csv",
  df_plan_v_tilde_full = NULL,
  fl_max_phaseout = 238000,
  it_bin_dollar_before_phaseout = 2500,
  fl_percheck_dollar = 100,
  fl_multiple = 58056,
  it_max_checks = 44,
  fl_tax_hh = 128580000,
  it_max_age = 64,
  it_min_age = 18,
  it_age_bins = 2,
  ar_svr_csv = c("age", "marital", "kids", "checks", "ymin", "mass", "survive",
    "vtilde", "ctilde"),
  ar_svr_groups = c("marital", "kids", "age_group", "ymin_group"),
  ar_svr_groups_stats = c("mass", "survive"),
  svr_checks = "checks",
  svr_v_value = "vtilde",
  svr_c_value = "ctilde",
  svr_mass = "mass",
  ar_rho = c(1),
  bl_threshold = FALSE,
  it_check_headorspouse = 12,
  it_check_perkids = 5,
  bl_given_firstcheck = FALSE,
  bl_non_inc_adjust = FALSE,
  bl_print = TRUE,
  bl_print_verbose = FALSE
)
}
\description{
Matlab Simulation Process for Allocation
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
